% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-.R
\name{tbl_}
\alias{tbl_}
\title{Wrap a function to support tibbles}
\usage{
tbl_(f, args = NULL)
}
\description{
Use this to make functions that only accept bare data frames
also accept tibbles.
This is a function constructor: it wraps a user-supplied function with logic
that converts inputs from tibbles to data frames.
}
\examples{
N <- 3
M <- 2

tbl <- tibble::tibble(
  g = gl(N, M),
  time = rep(seq_len(M), N),
  variable = rnorm(M * N)
)

try(reshape(
  tbl, v.names = "variable",
  idvar = "g", timevar = "time", direction = "wide"
))

tbl_(reshape)(
  tbl, v.names = "variable",
  idvar = "g", timevar = "time", direction = "wide"
)
}
